backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.2
django-filter>=23.5
openpyxl>=3.1
whitenoise>=6.5
dj-database-url>=2.1
